requirements.txt
sh
future
PyRIC
pyzmq
peakutils
numpy
scipy
git+https://github.com/cristianav/PyCRC.git
